MOve things around. Remove duplicated functions for stringify
