+++ geoh5py/shared/utils.py
@@ -39,6 +39,8 @@ import numpy as np
 from .exceptions import Geoh5FileClosedError
 
 
+UidOrNumeric = UUID | float | int | None
+StringOrNumeric = str | float | int
 # pylint: disable=too-many-lines
 
 if TYPE_CHECKING:
@@ -1439,3 +1441,10 @@ def map_to_class(
             class_map[identifier] = member
 
     return class_map
+
+
+def empty_string_to_none(value):
+    """Promote empty string to uid, and pass all other values."""
+    if value == "":
+        return None
+    return value
